rag-pipeline/requirements.txt
kagglehub>=0.2.0
pandas>=2.0.0
google-generativeai>=0.3.0
pinecone>=3.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

